Major restructuring of the way meaage tuples are sent on
The old 'writer' mechanism got broken up into three layers, a writer with
identical interface but only concerned with a specific marshalling method
(currently text or binary), a stream only concerned with how to send (net)
or store (file) marshalled data, and a queue between the two to manage the
potential differences in the writers production rate and the streamers
consumption rate.

The two new files 'text_writer.c' and 'bin_writer.c' a almost identical to the old
'file_writer.c' and 'net_writer.c' but instead of also dealing with files and
sockets, they obtain a 'mbuf' object from the queue (unfortunately named
'buffered_writer'), store marshalled data into it, and inform the queue of their
progress and obtain new 'mbuf' objects when they fill up.

The 'buffered_writer.c' object maintains a set of 'mbuf' objects, locks to
synchronize access, and a thread to actively push data into the associated
'stream'. The 'buffered_writer' will grow its internal buffer capacity to a
configurable max. size after which it starts dropping data. The currently
implemented strategy drops data from the end of the queue (the last mbuf).

The 'stream' objects implement a simple blocking 'write(chunk_of_bytes)'
interface which is called by the buffered_writer to empty the queue.

The 'file_stream', as the name suggests writes to a local file, while the
'net_stream' sends data over a network connection. The only new code added
disables SIGPIPE signals in the net_stream and have it attempt at a currently
hard-coded 10 sec interval to reconnect to the server and proeed on success.
Note: I believe that this is not necessarily enough to continue on the server
side, but I leave that to someone who is more familiar with the server's
restart capabilities.

That brings me to the short comings. The text protocol seems to work fine, but
binary one leads to 'incomplete message' errors on the server side. I didn't
follow up on these as I'm not familiar with the binary marshalling code.
The current design assumption in the buffered_writer is that an entire 'row'
is captured within an mbuf message, meaning that the msgptr is always at row
boundaries and we can safely throw away anything below the msgptr as long as
nothing has been read. Also everything up to the msgptr can be sent. Not sure
if this is relevant here.