Improvements to OML example application.
The example generator got extended with a -q flag to supress any printing
to stdout. Also the -s option was extended to suppress any delays in the
production loop if it is set to 0. These changes make the generator a good
stress tester for the buffered_writer's load shedding algorithm and the
multithreaded nature introduced by the 'buffered_writer'.