Add DBCLOSED hook command
When the last client for an experiment disconnects from the server, a
DBCLOSED event is sent to the event hook if present. Its one argument is
a URI with the full path to the database just closed. Currently only the
file: scheme is used, and this hook is only trigerred for SQLite3
backends.

An example script to push the data to iRODS is also provided.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>

fixes: 756