Hopefully fixed the reconnection issue. This fix adds a new MBuffer to buffered _writer which holds all the header/meta data information. This information is now also offered to the streams with each write message. The streams now need to keep track if the stream was just opened in which case they should first send the header data before anything else.
There are a few potential pitfalls. First, we will loose some data on the net writer as the OS will still accept some data into the socket even if the other side is no longer listening. When the stream reports that it has a problem and asks the buffered writer to call back in a while, the buffered writer is rewinding the associated chain's buffer to the beginning to start re-sending at message boundaries (which the stream doesn't know. I'm not a 100% confident that this is really working the way I believe it is. 

The other issue is that currently the net_stream is catching SIGPIPE signals. Copied the code from init.c but I'm not sure if that doesn't interfere with any signal handling the application is attempting to do. It seems to work, but I couldn't test it extensively.