Store poptGetNextOpt()'s return value appropriately
popt(3)'s poptGetNextOpt() returns a signed int, not a char. This signed
int is negative when no option, or an error, is found. Storing its
return value in a char loses this signedness in some case (big endian
machines). As the exit condition of the command-line parsing is that
poptGetNextOpt(...)>0, this led to endless loops.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>