Fix check failure when building ArchLinux package.
For some reason, after freeing the ClientHandler, the static string
exp_id becomes empty, and trying to reopen the database fails.

This fixes the symptom rather than the bug. This might suggest that
there is some memory dodginess happening in the ClientHandler, and that
finally freeing it overwrites some data in the stack which are not part
of the object.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>