Move sql_to_oml_type() out of lib/shared/schema.c
This is an SQLite3-specific function, so it has been moved into the
relevant adapter. Two functions of schema.c, schema_from_sql() and
schema_field_from_sql() have seen their prototype change to accomodate
the change: the DB adapter now needs to supply an
  OmlValueT (*reverse_typemap) (const char *s)
as their last argument.

Ultimately, this bit should be dropped, and a generic schema-storage in
the database's metadata should be used instead, as the PostgreSQL
backend currently does.

This commit also updates the documentation of related functions.

refs: #817, #837

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>