Resync after erroneous message in binary mode
Rather than disconnecting the client on invalid messages (bad schema, or
garbage), discard the current message with a warning and resync to the
next 0xAAAA, hoping for the best.

An extension would be to count the number of subsequent failures, or the
amount of invalid data, and eventually disconnect the client if it sends
too much garbage.

refs: #792