Filter out duplicate metadata in system tests
Upon (re)-connection, the client sends all its headers. Including the
very one it might be reconnecting to send. At this stage (following a
discussion with Max and Thierry), we err on the side of caution, and the
server doesn't do anything about duplicates; it just stores them again.

As the blobgen system test starts by sending metadata, that gets added
to the header, which gets sent first, then the actual metadata again.
This is not particularly nice and should probably be fixed later, but
tihs is not problematic. However, the server, not doing anything about
this, stores the metadata twice, which makes the test fail.

To fix this, we simply filter duplicates before the comparison.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>