Fix double-free bug in psql_adapter
With our current error-handling, we try to PQclear() PG results if they
are not NULL. However, in the process of the rest of the code, we
PQclear them as they become unneeded. We also need to set the pointer to
NULL, so the system doesn't try to clear them a second time. Another
option would be to keep them lingering until the end of the function,
and only clear them when we know we are about to overwrite them.

Also, expanded schema-parsing error message to indicate an issue with
per-2.10 databases.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>