Allow to override pgsql-inc
This is a fixup for the 2.10 release to build on Debian hosts.
It adds an optional pgsqlinc macro which, if defined, is added to the
./configure command line with the --pgsql-inc prefix.

As rpmbuild looks into various files for locally-defined macros, this
macro can be defined in, e.g., ~/.rpmmacros as
  %pgsqlinc /usr/include/postgresql
so the macro points to the right location on Debian.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>