Parse the PlantUML diagrams in Doxygen
They *need* to be introduced in Doxygen comments with

  @startuml{IMGNAME.png}
  ...
  @enduml

(i.e., not @startuml). The file extension is needed too, and it doesn't
seem possible to generate a different format (e.g., EPS) depending on
the Doxygen output format (HTML vs. PDF).

Also, when plantuml is detected, it includes the generated diagrams in
the distfile. For proper detection, the best is to create a small script
named plantuml somewhere in the PATH calling

  java -Djava.awt.headless=true -jar /PATH/TO/PLANTUML.jar $*

fixes: #1457

Signed-off-by: Olivier Mehani <shtrom@ssji.net>