Test the downstream connection before sending
Try to *read* the downstream connection from the server (even though we
don't expect anything from it) to check that it hasn't been closed
(i.e., a FIN has been sent, e.g., when the server dropped an idle
connection). If we read nothing (-EAGAIN), it's fine. Any other case
(0 data read, or error) means that something bad is happening, so we
mark the socket as disconnected and return to the writer function,
without wrongfully acknowledging the transfer of the last chunk, so it
is properly re-sent.

Also, don't skip unsent chunk when a disconnection occurs.
mbuf_reset_read() resets the pointer to the start of the message.  There
is an undefined semantic as to whether this is a read of write message.
and processChain assumes earlier that it is actually a *write* message.
Resetting it therefore doesn't achieve our objective of rewinding.
Moreover, we only skip over the Mbuffer at the end of a successfull
write to the socket, so we don't actually need to do anything before
that.

fixes: #1640

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>